refactor: split to new settings page
